+++ src/main.rs
@@ -88,10 +88,12 @@ fn exec(
         .map(|_| ())
 }
 
+const VERSION: &str = env!("CARGO_PKG_VERSION");
+
 fn main() -> ExitResult {
     env_logger::init();
 
-    let action = jump::prepare_boot().map_err(|e| {
+    let action = jump::prepare_boot(VERSION).map_err(|e| {
         Code::FAILURE.with_message(format!("Failed to prepare a scie jump action: {e}"))
     })?;
 
@@ -104,7 +106,7 @@ fn main() -> ExitResult {
         BootAction::Inspect((jump, lift)) => boot::inspect(jump, lift),
         BootAction::Install((scie, commands)) => boot::install(scie, commands),
         BootAction::List(commands) => boot::list(commands),
-        BootAction::Pack((jump, scie_jump_path)) => boot::pack(jump, scie_jump_path),
+        BootAction::Pack((jump, scie_jump_path)) => boot::pack(jump, scie_jump_path, VERSION),
         BootAction::Select(select_boot) => boot::select(select_boot),
         BootAction::Split((jump, lift, scie_path)) => boot::split(jump, lift, scie_path),
     }